GitRepoInspector: Deepen(): Instruct Git to fetch from the point of the deepest commit.
Without doing this, then Git may take commits from other branches, and never fetching our desired parent.

Note, I'm unable to replicate this in unit tests or integration tests without a lot of infrastructure. I can only repro it with a shallow HTTP clone.
It's possibly we can use `fetch --deepen=deltaDepth` instead of `fetch --depth=newDepth`, but doing so limits to Git remotes >= 2.11.0 (2016-11-23) instead of Git remotes >= 2.5.0 (2015-07-27) with `uploadpack.allowReachableSHA1InWant` set to `true`.

Because Git 2.5.0 is "brand new", according to Atlassian, and with it, fetching from a specific commit, a fallback is provided for servers running Git >=1.8.0 and <2.5.0.