GitRepoInspector: Trim the response from git incase whitespace chars present
When running in Visual Studio's MsBuild, `git cat-file` may respond with various whitespace around, so ensure we're resilient to it.

Note, the char array is used because .NET does not treat all unicode whitespace chars as "trimmable," and git, when running under Visual Studio's environment, returns a zero width non-breaking space that Verlite was previously not expecting.