Clone required git config options from the original repo
For now, only the http section will be copied. This is required for
GitHub Actions as authentication is done by adding a basic auth header
to outbound HTTP requests. This authentication header is set only at
a local level, not a user or system level, meaning the shadow repo
did not inherit authentication from the parent.

Fixes #141