Fixed a bug where min versions were not working correctly with post releases
The minimum version was not correctly factoring in the post releases by chopping
off the meta field when taking the core version. Now, a promotion to core version
is only allowed when: the min ver prerelease is not set; and the core version is
greater than the tag version.

In addition to this fix, prereleases are now correctly checked when a minimum version
contains a prerelease.